requirements.txt
pytest
qrcode
Pillow